Create an object of class SpatialPointsDataFrame

Set spatial coordinates and projection attributes to create an object of class SpatialPointsDataFrame from the observation ("observacao") table of one or more standardized datasets contained in the Free Brazilian Repository for Open Soil Data -- febr, http://www.ufsm.br/febr.

febr2sp(obj)

Arguments

obj
Object of class data.frame downloaded from febr using function observation.

Value

An object of class SpatialPointsDataFrame

Examples

# NOT RUN {
library(magrittr)
observation(dataset = "ctb0003", variable = "taxon",
            progress = FALSE, verbose = FALSE) %>%
  febr2sp() %>%
  spplot(zcol = "taxon_sibcs_2009", auto.key = list(columns = 3), scales = list(draw = TRUE)) +
  latticeExtra::layer(panel.grid(v = -1, h = -1))
# }