Go to febr

Go to one of the web pages of the Free Brazilian Repository for Open Soil Data -- febr --, including project and dataset web pages.

goto(dataset, table, page)

Arguments

dataset
Character vector indicating one dataset. The identification code should be as recorded in http://www.ufsm.br/febr/catalog/.

table
Character string indicating a table, i.e. the dataset table, "dataset", the observation table, "observacao", the layer table, "camada", or the metadata table, "metadado".

page
Character string indicating a web page of the febr. Options are: "febr", "view", "catalog", "search", "book", "package", "github", "forum", "unit", and "standard".

Examples

# Go to the main project page
goto(page = "febr")