Get metadata table

Download data from the metadata ("metadado") table of one or more datasets contained in the Free Brazilian Repository for Open Soil Data -- febr, http://www.ufsm.br/febr. This includes variable names, description of analytical methods, and identification of analysis laboratories.

metadata(dataset, progress = TRUE, verbose = TRUE)

Arguments

dataset
Character vector indicating one or more datasets. Identification codes should be as recorded in http://www.ufsm.br/febr/catalog/. Use dataset = "all" to download all datasets.

progress
(optional) Logical value indicating if a download progress bar should be displayed.

verbose
(optional) Logical value indicating if informative messages should be displayed. Generally useful to identify datasets with inconsistent data. Please report to febr-forum@googlegroups.com if you find any issue.

Value

A list of data frames or a data frame with metadata of the chosen dataset(s).

Examples

res <- metadata(dataset = "ctb0003")

#>   |                                                                              |                                                                      |   0%

#> 
#> Downloading dataset ctb0003...

#>   |                                                                              |======================================================================| 100%

head(res)

#>   dataset_id  tabela_id        campo_id                           campo_nome
#> 1    ctb0003 observacao   observacao_id          Identificação da observação
#> 2    ctb0003 observacao observacao_data                   Data da observação
#> 3    ctb0003 observacao   coord_sistema Sistema de referência de coordenadas
#> 4    ctb0003 observacao         coord_x                            Longitude
#> 5    ctb0003 observacao         coord_y                             Latitude
#> 6    ctb0003 observacao  coord_precisao             Precisão das coordenadas
#>                                                      campo_descricao
#> 1                     Identificador exclusivo da observação do solo.
#> 2                 Data da observação do solo no formato dia/mês/ano.
#> 3 Sistema de referência de coordenadas espaciais utilizado (WGS 84).
#> 4                        Coordenadas geográfica (latitude) em graus.
#> 5                       Coordenadas geográfica (longitude) em graus.
#> 6      Precisão com que as coordenadas espaciais foram determinadas.
#>   campo_fator campo_unidade campo_tipo metodo_id lab_nome lab_pais_id
#> 1          NA          <NA>      Texto       n/a      n/a         n/a
#> 2          NA          <NA>       Data       n/a      n/a         n/a
#> 3          NA          <NA>    Inteiro       n/a      n/a         n/a
#> 4           1             °       Real       n/a      n/a         n/a
#> 5           1             °       Real       n/a      n/a         n/a
#> 6           1             m    Inteiro       n/a      n/a         n/a
#>   lab_municipio_id lab_codigo_postal lab_rua_nome lab_rua_numero
#> 1              n/a               n/a          n/a            n/a
#> 2              n/a               n/a          n/a            n/a
#> 3              n/a               n/a          n/a            n/a
#> 4              n/a               n/a          n/a            n/a
#> 5              n/a               n/a          n/a            n/a
#> 6              n/a               n/a          n/a            n/a