Get standard table

Download data from the standard ("padrao") table of the Free Brazilian Repository for Open Soil Data -- febr, http://www.ufsm.br/febr. This includes codes and names of variable, measurement units, number of decimal places, type of data, and description of analytical methods. This is used to standardize the data contained in a dataset when downloading it via layer or observation.

standard(table, variable, unit, precision, expr)

Arguments

table
(optional) Character string indicating a table, i.e. the layer table, "camada", or the observation table, "observacao".

variable
(optional) Character vector indicating one or more variables. Accepts both specific identification codes, e.g. "ferro_oxalato_icpoes" and "carbono_cromo_30min150_mohr", as well as general identification codes, e.g. "ferro" and "carbono".

unit
(optional) Character vector indicating one or more measurement units. For example, "g/kg", "g/cm^3", and "cmolc/kg".

precision
(optional) Integer vector indicating one or more number of decimal places.

expr
(optional) Character string to be parsed and evaluated as a regular expression. For example, "campo_precisao > 0". Usage requires some knowledge of the structure of the standard table.

Value

A data frame with standards for selected variable(s).

References

Teixeira, P. C., Donagemma, G. K., Fontana, A., Teixeira, W. G. (2017) Manual de Métodos de Análise de Solo. Brasília: Embrapa.

See also

Examples

res <- standard(variable = "ferro")
head(res)

#>   tabela_id               campo_id              campo_nome
#> 1    camada  ferro_aquaregia_icpms             Ferro total
#> 2    camada ferro_aquaregia_icpoes             Ferro total
#> 3    camada   ferro_cloridrico_eaa             Ferro total
#> 4    camada    ferro_ditionito_eaa Ferro livre (extraível)
#> 5    camada    ferro_ditionito_eam Ferro livre (extraível)
#> 6    camada ferro_ditionito_icpoes Ferro livre (extraível)
#>                                                                                                                                                                                                                                                                                                                                                                                                                              campo_descricao
#> 1                                                                         Conteúdo de ferro total na fração terra fina seca ao ar (< 2 mm). Digestão à quente com ácido nítrico (HNO3) + ácido clorídrico (HCl) concentrados, numa proporção de 3:1 [aquaregia]. {ESPECIFICAR FORMA DE AQUECIMENTO, TEMPO E TEMPERATURA} Quantificação por espectrometria de massa com plasma acoplado indutivamente [icpms]. {INDICAR REFERÊNCIA DO MÉTODO}
#> 2                                                              Conteúdo de ferro total na fração terra fina seca ao ar (< 2 mm). Digestão à quente com ácido nítrico (HNO3) + ácido clorídrico (HCl) concentrados, numa proporção de 3:1 [aquaregia]. {ESPECIFICAR FORMA DE AQUECIMENTO, TEMPO E TEMPERATURA} Quantificação por espectrometria de emissão atômica com plasma acoplado indutivamente [icpoes]. {INDICAR REFERÊNCIA DO MÉTODO}
#> 3                                                                                                                                                           Conteúdo de ferro total na fração terra fina seca ao ar (< 2 mm). Digestão à quente com ácido clorídrico (HCl) 6 mol/L [cloridrico]. {ESPECIFICAR TEMPERATURA E TEMPO DE AQUECIMENTO} Quantificação por espectrometria de absorção atômica [eaa]. {INDICAR REFERÊNCIA DO MÉTODO}
#> 4                                     Conteúdo de ferro livre (extraível) na fração terra fina seca ao ar (< 2 mm). Extração com citrato de sódio (C6H5O7NA3.2H2O) 0,3 mol/L + bicarbonato de sódio (NaHCO3) 1 mol/L + ditionito de sódio [ditionito] sob aquecimento em banho-maria à temperatura de 75-80ºC. {ESPECIFICAR TEMPO DE AQUECIMENTO} Quantificação por espectrometria de absorção atômica [eaa]. {INDICAR REFERÊNCIA DO MÉTODO}
#> 5                Conteúdo de ferro livre (extraível) na fração terra fina seca ao ar (< 2 mm). Extração com citrato de sódio (C6H5O7NA3.2H2O) 0,3 mol/L + bicarbonato de sódio (NaHCO3) 1 mol/L + ditionito de sódio [ditionito] sob aquecimento em banho-maria à temperatura de 75-80ºC. {ESPECIFICAR TEMPO DE AQUECIMENTO} Quantificação por espectrofotometria de absorção molecular (colorimetria) [eam]. {INDICAR REFERÊNCIA DO MÉTODO}
#> 6 Conteúdo de ferro livre (extraível) na fração terra fina seca ao ar (< 2 mm). Extração com citrato de sódio (C6H5O7NA3.2H2O) 0,3 mol/L + bicarbonato de sódio (NaHCO3) 1 mol/L + ditionito de sódio [ditionito] sob aquecimento em banho-maria à temperatura de 75-80ºC. {ESPECIFICAR TEMPO DE AQUECIMENTO} Quantificação por espectrometria de emissão atômica com plasma acoplado indutivamente [icpoes]. {INDICAR REFERÊNCIA DO MÉTODO}
#>   campo_unidade campo_precisao campo_tipo campo_cat status
#> 1          g/kg              1       real   química      1
#> 2          g/kg              1       real   química      1
#> 3          g/kg              1       real   química      1
#> 4          g/kg              2       real   química      1
#> 5          g/kg              2       real   química      1
#> 6          g/kg              2       real   química      1