Get dataset table

Download data from the dataset ("dataset") table of one or more datasets contained in the Free Brazilian Repository for Open Soil Data -- febr, http://www.ufsm.br/febr. This includes dataset title and description, author and institution identification, dataset license, and much more.

dataset(dataset, progress = TRUE, verbose = TRUE, febr.repo = "remote")

Arguments

dataset
Character vector indicating one or more datasets. Identification codes should be as recorded in http://www.ufsm.br/febr/catalog/. Use dataset = "all" to download all datasets.

progress
(optional) Logical value indicating if a download progress bar should be displayed.

verbose
(optional) Logical value indicating if informative messages should be displayed. Generally useful to identify datasets with inconsistent data. Please report to febr-forum@googlegroups.com if you find any issue.

febr.repo
(optional) Character vector indicating where the data should be read. Defaults to febr.repo = "remote", i.e. the remote web server. Alternatively, a local directory path can be passed to febr.repo if the user has a local copy of the data repository.

Value

A list of data frames or a data frame with data of the chosen dataset(s).

Examples

res <- dataset(dataset = "ctb0003")

#>   |                                                                              |                                                                      |   0%

#> 
#> Downloading ctb0003-identificacao...

#>   |                                                                              |======================================================================| 100%