Get standard table

Download data from the standard ("padrao") table of the Free Brazilian Repository for Open Soil Data -- febr, http://www.ufsm.br/febr. This includes codes and names of variable, measurement units, number of decimal places, type of data, and description of analytical methods. This is used to standardize the data contained in a dataset when downloading it via layer or observation.

standard(table, variable, unit, precision, expr)

Arguments

table
(optional) Character string indicating a table, i.e. the layer table, "camada", or the observation table, "observacao".

variable
(optional) Character vector indicating one or more variables. Accepts both specific identification codes, e.g. "ferro_oxalato_icpoes" and "carbono_dicromato_30min150_mohr", as well as general identification codes, e.g. "ferro" and "carbono".

unit
(optional) Character vector indicating one or more measurement units. For example, "g/kg", "g/cm^3", and "cmolc/kg".

precision
(optional) Integer vector indicating one or more number of decimal places.

expr
(optional) Character string to be parsed and evaluated as a regular expression. For example, "campo_precisao > 0". Usage requires some knowledge of the structure of the standard table.

Value

A data frame with standards for selected variable(s).

References

Teixeira, P. C., Donagemma, G. K., Fontana, A., Teixeira, W. G. (2017) Manual de Métodos de Análise de Solo. Brasília: Embrapa.

See also

The standard table at https://goo.gl/hi77sB

Examples

# NOT RUN {
a <- standard(variable = "ferro")
# }